requirements.txt
# -----------------------------
# Environment
# -----------------------------
python-dotenv>=1.0.0

# -----------------------------
# LangChain + LangGraph
# -----------------------------
langchain>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
langgraph>=0.0.56

# -----------------------------
# Vector Store (Pinecone)
# -----------------------------
pinecone-client>=5.1.0

# -----------------------------
# OpenAI + Token Utilities
# -----------------------------
langchain-openai>=0.2.0
openai>=1.30.0
tiktoken>=0.6.0

# -----------------------------
# Validation
# -----------------------------
pydantic>=2.6.0,<3.0.0

# -----------------------------
# File Parsing
# -----------------------------
python-docx>=1.1.0

# -----------------------------
# Database
# -----------------------------
pyodbc>=5.0.1

# -----------------------------
# Logging
# -----------------------------
logtail-python>=0.3.0

# -----------------------------
# Utilities
# -----------------------------
uuid6>=2024.1
sentence-transformers
rich>=13.7.0
langchain-huggingface>=0.1.2
